Rename js folder to logic, which makes more sense
